+++ src/main/java/xyz/nikitacartes/easyauth/config/MainConfigV1.java
@@ -24,6 +24,7 @@ public class MainConfigV1 extends ConfigTemplate {
     public boolean enableGlobalPassword = false;
     public boolean singleUseGlobalPassword = false;
     public boolean hidePlayerCoords = false;
+    public boolean vanishUntilAuth = true;
     public boolean debug = false;
     public long configVersion = 2;
     public WorldSpawn worldSpawn = new WorldSpawn();
@@ -62,6 +63,7 @@ public class MainConfigV1 extends ConfigTemplate {
         configValues.put("enableGlobalPassword", wrapIfNecessary(enableGlobalPassword));
         configValues.put("singleUseGlobalPassword", wrapIfNecessary(singleUseGlobalPassword));
         configValues.put("hidePlayerCoords", wrapIfNecessary(hidePlayerCoords));
+        configValues.put("vanishUntilAuth", wrapIfNecessary(vanishUntilAuth));
         configValues.put("worldSpawn.dimension", wrapIfNecessary(worldSpawn.dimension));
         configValues.put("worldSpawn.x", wrapIfNecessary(worldSpawn.x));
         configValues.put("worldSpawn.y", wrapIfNecessary(worldSpawn.y));