+++ src/main/java/xyz/nikitacartes/easyauth/event/AuthEventHandler.java
@@ -15,6 +15,7 @@ import net.minecraft.text.Text;
 import net.minecraft.util.ActionResult;
 import net.minecraft.util.Uuids;
 import net.minecraft.util.math.BlockPos;
+import xyz.nikitacartes.easyauth.integrations.VanishIntegration;
 import xyz.nikitacartes.easyauth.storage.PlayerEntryV1;
 import xyz.nikitacartes.easyauth.utils.FloodgateApiHelper;
 import xyz.nikitacartes.easyauth.utils.PlayerAuth;
@@ -123,6 +124,11 @@ public class AuthEventHandler {
         if (extendedConfig.skipAllAuthChecks) {
             playerAuth.easyAuth$setAuthenticated(true);
         }
+
+        if (config.vanishUntilAuth && technicalConfig.vanishLoaded) {
+            ((PlayerAuth) player).easyAuth$wasVanished(VanishIntegration.isVanished(player));
+            VanishIntegration.setVanished(player, true);
+        }
     }
 
     // Player joining the server
@@ -163,7 +169,9 @@ public class AuthEventHandler {
             PlayerEntryV1 playerCache = playerAuth.easyAuth$getPlayerEntryV1();
             playerCache.lastAuthenticatedDate = ZonedDateTime.now();
             playerCache.update();
-        } else if (config.hidePlayerCoords) {
+            return;
+        }
+        if (config.hidePlayerCoords) {
             ((PlayerAuth) player).easyAuth$restoreTrueLocation();
         }
     }